+++ src/lib.rs
@@ -7,3 +7,5 @@ pub mod simulatable_message;
 pub mod simulate_data;
 
 pub mod imd_poll;
+
+pub mod can;
